+++ app/src/main/java/nextstep/signup/MainActivity.kt
@@ -7,12 +7,11 @@ import androidx.activity.compose.setContent
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Surface
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import nextstep.signup.ui.screen.SignupScreen
+import nextstep.signup.ui.screen.signup.SignupScreen
 import nextstep.signup.ui.theme.SignupTheme
 
 @SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
