+++ app/src/main/java/nextstep/signup/ui/component/SignupTextField.kt
@@ -1,4 +1,4 @@
-package nextstep.signup.ui.component.textfield
+package nextstep.signup.ui.component
 
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material3.MaterialTheme
