+++ app/src/main/java/nextstep/signup/ui/screen/signup/component/EmailTextField.kt
@@ -1,31 +1,29 @@
-package nextstep.signup.ui.component.textfield
+package nextstep.signup.ui.screen.signup.component
 
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.derivedStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.KeyboardType
 import nextstep.signup.R
 import nextstep.signup.ui.ValidationState
+import nextstep.signup.ui.component.SignupTextField
 
 @Composable
-fun PasswordTextField(
-    password: String,
-    onPasswordChange: (String) -> Unit,
+fun EmailTextField(
+    email: String,
+    onEmailChange: (String) -> Unit,
     validationState: ValidationState,
     modifier: Modifier = Modifier
 ) {
     SignupTextField(
-        value = password,
-        onValueChange = onPasswordChange,
-        label = stringResource(R.string.password),
+        value = email,
+        onValueChange = onEmailChange,
+        label = stringResource(R.string.email),
         keyboardOptions = KeyboardOptions(
             autoCorrectEnabled = false,
-            keyboardType = KeyboardType.Password
+            keyboardType = KeyboardType.Email
         ),
-        secure = true,
         validationState = validationState,
         modifier = modifier
     )
