+++ app/src/main/java/nextstep/signup/ui/screen/signup/component/PasswordConfirmTextField.kt
@@ -1,4 +1,4 @@
-package nextstep.signup.ui.component.textfield
+package nextstep.signup.ui.screen.signup.component
 
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.runtime.Composable
@@ -7,6 +7,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.KeyboardType
 import nextstep.signup.R
 import nextstep.signup.ui.ValidationState
+import nextstep.signup.ui.component.SignupTextField
 
 @Composable
 fun PasswordConfirmTextField(
