+++ app/src/main/java/nextstep/signup/ui/screen/signup/component/PasswordTextField.kt
@@ -1,32 +1,30 @@
-package nextstep.signup.ui.component.textfield
+package nextstep.signup.ui.screen.signup.component
 
-import android.util.Log
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.derivedStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.KeyboardType
 import nextstep.signup.R
 import nextstep.signup.ui.ValidationState
-import nextstep.signup.ui.validateEmail
+import nextstep.signup.ui.component.SignupTextField
 
 @Composable
-fun EmailTextField(
-    email: String,
-    onEmailChange: (String) -> Unit,
+fun PasswordTextField(
+    password: String,
+    onPasswordChange: (String) -> Unit,
     validationState: ValidationState,
     modifier: Modifier = Modifier
 ) {
     SignupTextField(
-        value = email,
-        onValueChange = onEmailChange,
-        label = stringResource(R.string.email),
+        value = password,
+        onValueChange = onPasswordChange,
+        label = stringResource(R.string.password),
         keyboardOptions = KeyboardOptions(
             autoCorrectEnabled = false,
-            keyboardType = KeyboardType.Email
+            keyboardType = KeyboardType.Password
         ),
+        secure = true,
         validationState = validationState,
         modifier = modifier
     )
